試合速報の処理をScoreReportFragmentに切り出す
