--- app/src/main/java/com/example/sasata299/carpqrapplication/Fragment1.java
@@ -1,22 +1,0 @@
-package com.example.sasata299.carpqrapplication;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-/**
- * Created by sasata299 on 2016/11/04.
- */
-
-public class Fragment1 extends Fragment {
-    public Fragment1() {
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment1, container, false);
-        return view;
-    }
-}
